backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.5
pydantic-settings==2.8.1
python-dotenv==1.0.1
boto3>=1.35.0
aiobotocore>=2.19.0
python-multipart==0.0.20 